fit tests to testing-guideline
